Refactor loan amount type from Integer to Long for consistency and accuracy

+++ backend/src/main/java/org/example/backend/constants/LoanConstants.java
@@ -5,8 +5,8 @@ package org.example.backend.constants;
  * Holds all constants in  place that are needed for application.
  */
 public class LoanConstants {
-    public static final int MINIMUM_LOAN_AMOUNT = 2000;
-    public static final int MAXIMUM_LOAN_AMOUNT = 10000;
+    public static final long MINIMUM_LOAN_AMOUNT = 2000;
+    public static final long MAXIMUM_LOAN_AMOUNT = 10000;
     public static final int MINIMUM_LOAN_PERIOD = 12;
     public static final int MAXIMUM_LOAN_PERIOD = 60;
     public static final double MINIMUM_CREDIT_SCORE = 1.0;
